Changes to remove _build_coordinate from extract_data.py
